requirements.txt
streamlit==1.32.0
scikit-learn==1.4.1.post1
joblib==1.3.2
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.26.4
